+++ src/raycast.c
@@ -6,10 +6,9 @@
 /*   By: flafi <flafi@student.42heilbronn.de>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 01:51:03 by flafi             #+#    #+#             */
-/*   Updated: 2024/04/02 01:51:04 by flafi            ###   ########.fr       */
+/*   Updated: 2024/04/04 01:00:56 by flafi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "../include/cub_3d.h"
